Integrate ore scanner into project system
