+++ index.html
@@ -54,7 +54,6 @@
     <script src="src/js/structuresUI.js"></script>
     <script src="src/js/day-night-cycle.js"></script>
     <script src="src/js/day-night-setting.js"></script>
-    <script src="src/js/ore-scanning.js"></script>
     <script src="src/js/advanced-research/self-replicating-ships.js"></script>
     <script src="src/js/advanced-research/hive-mind-androids.js"></script>
     <script src="src/js/resource.js"></script>
