+++ tests/oreScanningEffectable.test.js
@@ -1,7 +1,8 @@
 const EffectableEntity = require('../src/js/effectable-entity.js');
 // expose globally so OreScanning can extend it when required
 global.EffectableEntity = EffectableEntity;
-const OreScanning = require('../src/js/ore-scanning.js');
+global.Project = class extends EffectableEntity {};
+const ScannerProject = require('../src/js/projects/ScannerProject.js');
 
 describe('OreScanning integration with EffectableEntity', () => {
   const params = {
@@ -13,7 +14,8 @@ describe('OreScanning integration with EffectableEntity', () => {
   };
 
   test('instance inherits addEffect method', () => {
-    const scanner = new OreScanning(params);
+    const scanner = new ScannerProject({}, 'test');
+    scanner.initializeScanner(params);
     expect(typeof scanner.addEffect).toBe('function');
   });
 });
